Side Project 1.1
This is for Khan academy because I felt bored doing midpoint formulas and finding the second point in a line segment after being given the first and midpoint.

The first one works by finding the average distance between both x's and both y's

The second one works because the formula to find the second point given the first one and the midpoint:
(-4, 8) , (1, 7.5)​
​​1/2(−4+x)=1

only actually requires someone to multiply the 1 (representing the midpoint's x value) times 2 (canceling the 1/2) and then subtracting the -4 (representing the first point's x value) from the midpoint's x value *2. This works becasue subtracting a negative results in a positive, so if the number is negative, it adds, if positive, it subtracts, thereby moving the number over to the other side even if it is negative. This is how to find the value of x, just replacing the y values into it, one finds the y values. You could also just use find x for both x and y and scrap the find y function for both of them, reducing it to one definition and two operations instead of two definitions and two operations.